Tutto corretto, testiamo.

+++ ft_putnbr_u.c
@@ -4,6 +4,6 @@ int	ft_putnbr_u(va_list x)
 {
 	unsigned int	n;
 
-	n = (unsigned int)va_arg(x, int);
+	n = va_arg(x, unsigned int);
 	return (ft_print_int_recursive(n));
 }
